+++ src/MigrateSource/ArraySource.php
@@ -10,54 +10,21 @@ use Shuttle\SourceItem;
  * Class ArraySource
  * @package ShuttleTest\Fixture
  */
-class ArraySource implements SourceInterface
+class ArraySource extends CallbackSource
 {
     /**
-     * @var array
-     */
-    private $items;
-
-    /**
      * ArraySource constructor.
-     * @param iterable $items
+     * @param iterable|array|\Traversable $items  Keys/indicies must be IDs
      */
     public function __construct(iterable $items)
     {
-        $this->items = ($items instanceof \Traversable) ? iterator_to_array($items) : $items;
-    }
+        $getItemsCallback = function() use ($items) {
+            foreach ($items as $itemId => $itemData) {
+                $out[$itemId] = ($itemData instanceOf SourceItem) ? $itemData : new SourceItem($itemId, $itemData);
+            }
+            return $out ?? [];
+        };
 
-    /**
-     * If is countable, return the number of source items, or NULL if unknown
-     * @return int|null
-     */
-    public function countSourceItems(): ?int
-    {
-        return count($this->items);
-    }
-
-    /**
-     * Get the next source record, represented as an array
-     *
-     * Return an array for the next item, or NULL for no more item
-     *
-     * @return iterable|SourceItem[]
-     */
-    public function getSourceIdIterator(): iterable
-    {
-        return array_keys($this->items);
-    }
-
-    /**
-     * @param string $id
-     * @return SourceItem
-     * @throws MissingItemException  If source item is not found
-     */
-    public function getSourceItem(string $id): SourceItem
-    {
-        if (array_key_exists($id, $this->items)) {
-            return new SourceItem($id, $this->items[$id]);
-        } else {
-            throw new MissingItemException('Missing Item: ' . $id);
-        }
+        parent::__construct($getItemsCallback, true);
     }
 }